many to many  completed

+++ src/main/java/org/lessons/java/springlamiapizzeriacrud/controller/IngredientController.java
@@ -2,12 +2,15 @@ package org.lessons.java.springlamiapizzeriacrud.controller;
 
 import jakarta.validation.Valid;
 import org.lessons.java.springlamiapizzeriacrud.model.Ingredient;
+import org.lessons.java.springlamiapizzeriacrud.model.Pizza;
 import org.lessons.java.springlamiapizzeriacrud.repository.IngredientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -62,6 +65,17 @@ public class IngredientController {
 
     @PostMapping("/delete/{id}")
     public String delete(@PathVariable Integer id) {
+        //prima di eliminare ingrediente lo dissocio da tutte le pizze
+        Optional<Ingredient> result = ingredientRepository.findById(id);
+        if (result.isEmpty()) {
+            throw new ResponseStatusException((HttpStatus.NOT_FOUND));
+        }
+        //ingrediente da eliminare
+        Ingredient ingredientToDelete = result.get();
+        //per ogni pizza associata all' ingrediente da eliminare
+        for (Pizza pizza : ingredientToDelete.getPizza()) {
+            pizza.getIngredients().remove(ingredientToDelete);
+        }
         ingredientRepository.deleteById(id);
         return "redirect:/ingredients";
     }
